CachedThreadPool directly extends ThreadPoolExecutor.
